==> Data Cleaning/1_Filtering CPC codes.ipynb <==
# Filtering by CPC codes

### Assessing total number of unique patents in PATSAT

import pandas as pd
fields = ['appln_id']
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls201_part01.csv',skipinitialspace=True, header=0,usecols=fields)
df1['appln_id'].nunique()
#50,000,000

import pandas as pd
fields = ['appln_id']
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls201_part02.csv',skipinitialspace=True, header=0,usecols=fields)
df2['appln_id'].nunique()
#50,000,000

import pandas as pd
fields = ['appln_id']
df3 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls201_part03.csv',skipinitialspace=True, header=0,usecols=fields)
df3['appln_id'].nunique()
#5,355,385

### Total number of unique patents classified under CPC

import pandas as pd
fields = ['appln_id', 'cpc_class_symbol']
df4 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls224_part01.csv',skipinitialspace=True, header=0,usecols=fields)
df4['appln_id'].nunique() #number of unique application ids
# 38,554,358

df5 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls224_part02.csv',skipinitialspace=True, header=0,usecols=fields)
df5['appln_id'].nunique()
# 18,320,029

The database has ~105 million unique application ids, of which ~57 million have also been classified under CPC. 

# Filtering by CPC code

### Filtering to CPC1.csv

import pandas as pd
fields = ['appln_id', 'cpc_class_symbol']
filter_list = ['Y02A','Y02B','Y02C','Y02D','Y02E','Y02P','Y02T','Y02W','Y04S','Y10S','Y10T']
df6 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls224_part01.csv',skipinitialspace=True, chunksize=5000, header=0,usecols=fields)
header = True
for chunk in df6:
    chunk['cpc_class_symbol'] = chunk['cpc_class_symbol'].str.replace(" ","") #remove spaces 
    validid = chunk['cpc_class_symbol'].str.contains('Y02A|Y02B|Y02C|Y02D|Y02E|Y02P|Y02T|Y02W|Y04S|Y10S|Y10T', regex=True)
    chunk = chunk[validid]
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/CPCcodes_1.csv', index=False, mode='a', header=header)
    header = False
#     print(df3.shape)
#     print(df3.dtypes)
#     print(df3.isnull().sum()) #check for missing values
#     print(df3.head())

import pandas as pd
fields = ['appln_id']
df7 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPCcodes_1.csv',skipinitialspace=True, header=0,usecols=fields)
df7['appln_id'].nunique()
#6,136,206

### Filtering to CPC2.csv

import pandas as pd
fields = ['appln_id', 'cpc_class_symbol']
filter_list = ['Y02A','Y02B','Y02C','Y02D','Y02E','Y02P','Y02T','Y02W','Y04S','Y10S','Y10T']
df8 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls224_part02.csv',skipinitialspace=True, chunksize=5000, header=0,usecols=fields)
header = True
for chunk in df8:
    chunk['cpc_class_symbol'] = chunk['cpc_class_symbol'].str.replace(" ","") #remove spaces 
    validid = chunk['cpc_class_symbol'].str.contains('Y02A|Y02B|Y02C|Y02D|Y02E|Y02P|Y02T|Y02W|Y04S|Y10S|Y10T', regex=True)
    chunk = chunk[validid]
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPCcodes_2.csv', index=False, mode='a', header=header)
    header = False

import pandas as pd
fields = ['appln_id']
df9 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPCcodes_2.csv',skipinitialspace=True, header=0,usecols=fields)
df9['appln_id'].nunique()
#2,414,168

~8.5 million Y02 patents
