==> Data Cleaning/2_Assigning Patent Labels.ipynb <==
import pandas as pd
import numpy as np
labels = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPClabels.csv',skipinitialspace=True)

#Label Y02
def cpc_symbol_to_category(df_column, symbol, category):
    current_dict = dict.fromkeys(df_column[df_column.str.contains(symbol)], category) #build dictionary
    return df_column.replace(current_dict) #replace keys with values
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/CPCcodes_1.csv',skipinitialspace=True, chunksize=5000, header=0)
extra_labels = pd.DataFrame([['Y02E30', 'Nuclear'],['Y02E50', 'Non-Fossil'], ['Y02E60','Enable'], ['Y02E', 'Other-6']], columns=labels.columns.values)

header = True
for chunk in df1:
    chunk_modified = chunk.copy()
    chunk_modified['Category'] = chunk_modified['cpc_class_symbol'] #create a dummy column to be replaced later
    for index, row in labels.iterrows():
        chunk_modified['Category'] = cpc_symbol_to_category(chunk_modified['Category'], row['cpc_class_symbol'], row['Category'])
    for index, row in extra_labels.iterrows():
        chunk_modified['Category'] = cpc_symbol_to_category(chunk_modified['Category'], row['cpc_class_symbol'], row['Category'])
    chunk_modified.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPC_with_labels.csv', index=False, mode='a', header=header)
    header = False

#Label Y02
def cpc_symbol_to_category(df_column, symbol, category):
    current_dict = dict.fromkeys(df_column[df_column.str.contains(symbol)], category)
    return df_column.replace(current_dict)
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPCcodes_2.csv',skipinitialspace=True, chunksize=5000, header=0)
extra_labels = pd.DataFrame([['Y02E30', 'Nuclear'],['Y02E50', 'Non-Fossil'], ['Y02E60','Enable'], ['Y02E', 'Other-6']], columns=labels.columns.values)

header = True
for chunk in df2:
    chunk_modified = chunk.copy()
    chunk_modified['Category'] = chunk_modified['cpc_class_symbol']
    for index, row in labels.iterrows():
        chunk_modified['Category'] = cpc_symbol_to_category(chunk_modified['Category'], row['cpc_class_symbol'], row['Category'])
    for index, row in extra_labels.iterrows():
        chunk_modified['Category'] = cpc_symbol_to_category(chunk_modified['Category'], row['cpc_class_symbol'], row['Category'])
    chunk_modified.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPC_with_labels_2.csv', index=False, mode='a', header=header)
    header = False

import pandas as pd
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPC_with_labels.csv',skipinitialspace=True, header=0)
df1['appln_id'].nunique()
# 6,136,206

df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPC_with_labels_2.csv',skipinitialspace=True, header=0)
df2['appln_id'].nunique()
# 2,414,168

We retain all 8.5 million patents
