==> Data Cleaning/4_Extracting country of residence.ipynb <==
import pandas as pd
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls207_part01.csv',skipinitialspace=True, header=0)
ids = df1['appln_id']
ids1 = pd.DataFrame(data = ids)
df1['appln_id'].nunique()
#84,076,043

ids1

df1['person_id'].nunique()
#61,558,031

~84 million unique applications have persons associated with them, and ~62 million unique persons have contributed to the dataset

#checking how many of the 8.5 million Y02 patents have associated person ids
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPC_with_labels.csv', chunksize=500000, skipinitialspace=True, header=0)
header = True
chunkid=0
for chunk in df2:
    chunk = ids1.merge(chunk, on='appln_id', how='inner') #filter by ids
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Check1.csv', index=False, mode='a', header=header)
    print (chunkid)
    chunkid = chunkid + 1
    header = False
df3 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Check1.csv', skipinitialspace=True, header=0)
df3['appln_id'].nunique()
#5,758,923

df3

df4 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPC_with_labels_2.csv', chunksize=500000, skipinitialspace=True, header=0)
header = True
chunkid=0
for chunk in df4:
    chunk = ids1.merge(chunk, on='appln_id', how='inner') #filter by ids
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Check2.csv', index=False, mode='a', header=header)
    print (chunkid)
    chunkid = chunkid + 1
    header = False
df5 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Check2.csv', skipinitialspace=True, header=0)
df5['appln_id'].nunique()
#2,283,814

del df1, df2, df3, df4, df5, ids

~8 million Y02 patents have persons associated with them

### Filter Inventors by Application id

First I extract all relevant Person Id's from the larger database

import pandas as pd
import numpy as np
field = ['appln_id']
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPCcodes_1.csv',skipinitialspace=True, usecols = field) 
ids = df2.drop_duplicates() #extract relevant unique ids

fields = ['person_id', 'appln_id']
df3 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls207_part01.csv',skipinitialspace=True, chunksize=200000, header=0,usecols=fields)
header = True
chunkid=0
for chunk in df3:
    chunk = ids.merge(chunk, on='appln_id', how='inner') #filter by ids
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Persons_1.csv', index=False, mode='a', header=header)
    print (chunkid)
    chunkid = chunkid + 1
    header = False

del ids, df2
import pandas as pd
import numpy as np
field = ['appln_id']
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CPCcodes_2.csv',skipinitialspace=True, usecols = field) 
ids = df2.drop_duplicates() #extract relevant unique ids

fields = ['person_id', 'appln_id']
df3 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls207_part01.csv',skipinitialspace=True, chunksize=500000, header=0,usecols=fields)
header = True
chunkid=0
for chunk in df3:
    chunk = ids.merge(chunk, on='appln_id', how='inner') #filter by ids
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles1/Persons_2.csv', index=False, mode='a', header=header)
    print (chunkid)
    chunkid = chunkid + 1
    header = False

del ids, df2

### Extract Countries of Residence

Then I use Person id's to filter out country codes

import pandas as pd
import numpy as np
#extract relevant person ids
field = ['person_id']
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Persons_1.csv',skipinitialspace=True, usecols = field) 
ids = df1.drop_duplicates() #extract relevant unique person ids

fields = ['person_id', 'person_ctry_code']
#once again, parse through relevant person ids and residence country code columns
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls226_part01.csv',skipinitialspace=True, chunksize=500000, header=0,usecols=fields)
header = True
chunkid=0
for chunk in df2:
    chunk = ids.merge(chunk, on='person_id', how='inner') #filter by person ids
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Residence_1.csv', index=False, mode='a', header=header)
    print (chunkid)
    chunkid = chunkid + 1
    header = False

del df1, df2, ids

#similarly for Persons 2
import pandas as pd
import numpy as np
field = ['person_id']
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Persons_2.csv',skipinitialspace=True, usecols = field) 
ids = df1.drop_duplicates() #extract relevant unique person ids
import pandas as pd
import numpy as np
fields = ['person_id', 'person_ctry_code']
#once again, parse through relevant person ids and residence country code columns
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls226_part02.csv',skipinitialspace=True, chunksize=500000, header=0,usecols=fields)
header = True
chunkid=0
for chunk in df2:
    chunk = ids.merge(chunk, on='person_id', how='inner') #filter by person ids
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Residence_2.csv', index=False, mode='a', header=header)
    print (chunkid)
    chunkid = chunkid + 1
    header = False

del df2
import pandas as pd
import numpy as np
#once again, parse through relevant person ids and residence country code columns
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/tls226_part02.csv',skipinitialspace=True, header=0, nrows=100)
df2.head()

del df1, df2, ids

#merging names and country codes
import pandas as pd
import numpy as np
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Persons_1.csv',skipinitialspace=True, header=0)
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/Residence_1.csv',skipinitialspace=True, header=0)
df3 = pd.merge(df1, df2.drop_duplicates() , on="person_id", how = 'inner')

df3.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CountryPersons1.csv', encoding='utf-8', index=False)

import pandas as pd
field = ['appln_id']
df4 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CountryPersons1.csv', skipinitialspace=True, usecols=field, header=0)
df4['appln_id'].nunique()
#5,701,090

del df1, df2, df3, df4

import pandas as pd
import numpy as np
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Persons_2.csv',skipinitialspace=True, header=0)
df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Residence_2.csv',skipinitialspace=True, header=0)
df3 = pd.merge(df1, df2.drop_duplicates() , on="person_id", how = 'inner')

df3.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/CountryPersons2.csv', encoding='utf-8', index=False)

import pandas as pd
field = ['appln_id']
df4 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CountryPersons2.csv', skipinitialspace=True, usecols=field, header=0)
df4['appln_id'].nunique()
#960,054

del df1, df2, df3, df4

There are 6,661,144 unique patents with non-overlapping inventors.

### Divide person ids and country codes based on application categories

import pandas as pd
import numpy as np
field = ['appln_id']
#first try with one category
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Adapt.csv',skipinitialspace=True, usecols = field) 
ids = df1.drop_duplicates() #extract relevant unique ids

df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CountryPersons1.csv',skipinitialspace=True, chunksize=500000, header=0)
header = True
for chunk in df2:
    if (chunk['appln_id'].dtype is not ids['appln_id'].dtype):
        chunk['appln_id'] = pd.to_numeric(chunk['appln_id'], errors='coerce')
        chunk.dropna(subset=['appln_id'], inplace=True)
        chunk['appln_id'] = chunk['appln_id'].astype(ids['appln_id'].dtype)
    chunk = ids.merge(chunk, on='appln_id', how='inner') #filter by ids
    chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Extras/Adapt_residence.csv', index=False, mode='a', header=header)
    header = False

del df1, df2

#creating a for loop for the remaining categories
import pandas as pd
import numpy as np
Category = ['Build', 'Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Other-1', 'Other-2', 'Other-3', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Other-4', 'Other-5', 'Nuclear', 'Non-Fossil', 'Enable','Other-6']
field = ['appln_id']
for c in Category:
    df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/'+ c +'.csv',skipinitialspace=True, usecols = field) 
    ids = df1.drop_duplicates() #extract relevant unique ids
    df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/General/CountryPersons1.csv',skipinitialspace=True, chunksize=500000, header=0)
    header = True
    for chunk in df2:
        if (chunk['appln_id'].dtype is not ids['appln_id'].dtype):
            chunk['appln_id'] = pd.to_numeric(chunk['appln_id'], errors='coerce')
            chunk.dropna(subset=['appln_id'], inplace=True)
            chunk['appln_id'] = chunk['appln_id'].astype(ids['appln_id'].dtype)
        chunk = ids.merge(chunk, on='appln_id', how='inner') #filter by ids
        chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Extras/' + c + '_residence.csv', index=False, mode='a', header=header)
        header = False
    del df1, df2

Similarly for the second half of the data

import pandas as pd
import numpy as np
Category = ['Adapt','Build', 'Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Other-1', 'Other-2', 'Other-3', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Other-4', 'Other-5', 'Nuclear', 'Non-Fossil', 'Enable','Other-6']
field = ['appln_id']
for c in Category:
    df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/'+ c +'.csv',skipinitialspace=True, usecols = field) 
    ids = df1.drop_duplicates() #extract relevant unique ids
    df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/CountryPersons2.csv',skipinitialspace=True, chunksize=500000, header=0)
    header = True
    for chunk in df2:
        if (chunk['appln_id'].dtype is not ids['appln_id'].dtype):
            chunk['appln_id'] = pd.to_numeric(chunk['appln_id'], errors='coerce')
            chunk.dropna(subset=['appln_id'], inplace=True)
            chunk['appln_id'] = chunk['appln_id'].astype(ids['appln_id'].dtype)
        chunk = ids.merge(chunk, on='appln_id', how='inner') #filter by ids
        chunk.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Extras/' + c + '_residence2.csv', index=False, mode='a', header=header)
        header = False
    del df1, df2

import pandas as pd
import numpy as np
#merging tables
Category = ['Adapt','Build', 'Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Other-1', 'Other-2', 'Other-3', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Other-4', 'Other-5', 'Nuclear', 'Non-Fossil', 'Enable','Other-6']
for c in Category:
    df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/'+ c + '_residence.csv',skipinitialspace=True) 
    df2 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/'+ c + '_residence2.csv',skipinitialspace=True)
    df3 = pd.concat([df1, df2])
    df3.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Extras/'+ c + '_Residence_complete.csv', encoding='utf-8', index=False)
    del df1, df2, df3

Check maximum investors per record

import pandas as pd
import numpy as np
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Persons_1.csv',skipinitialspace=True, header=0)
items_counts = df1['appln_id'].value_counts()
max_item = items_counts.max()
del df1
#There are a maximum of 132 inventors associated with a patent

import pandas as pd
import numpy as np
df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Persons_2.csv',skipinitialspace=True, header=0)
items_counts = df1['appln_id'].value_counts()
max_item = items_counts.max()
del df1
#There are a maximum of 134 inventors associated with a patent

df1['appln_id'].value_counts().idxmax()

df1.loc[df1['appln_id'] == 478696789]


Therefore we will randmly pick 5 inventors per application

df2 = df1.groupby('appln_id', group_keys=False).apply(lambda x: x.sample(min(len(x), 5)))

items_counts = df2['appln_id'].value_counts()
max_item = items_counts.max()
max_item

del df1, df2

So we use this function and loop over all categories

import pandas as pd
import numpy as np
import time
Category = ['Adapt','Build', 'Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Other-1', 'Other-2', 'Other-3', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Other-4', 'Other-5', 'Nuclear', 'Non-Fossil', 'Enable','Other-6']
for c in Category:
    df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/'+ c + '_Residence_complete.csv', skipinitialspace=True, chunksize=2000000, header=0)
    header = True
    iteration = 0 
    for chunk in df1:
        print(c, iteration)
        iteration += 1
        time1 = time.time()
        df2 = chunk.groupby('appln_id', group_keys=False).apply(lambda x: x.sample(min(len(x), 5)))
        time2 = time.time()
        print(time2 - time1)
        df3 = df2.groupby('appln_id', group_keys=False).apply(lambda x: x.assign(inventor_number=np.arange(min(len(x),5))))
        time3 = time.time()
        print(time3 - time2)
        df4 = pd.pivot(df3, values=['person_id', 'person_ctry_code'],
                        index=['appln_id'],
                        columns=['inventor_number'])
        df4['appln_id'] = df4.index.values
        time4 = time.time()
        print(time4 - time3)
        df4.to_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Extras/'+ c + '_Residence_concise.csv', encoding='utf-8', index=False, mode='a', header=header)
        header = False
    del df1, df2

import pandas as pd
df3 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles1/Transport_Residence_concise.csv', skipinitialspace=True,nrows=10, header=0)
df3.head()

import pandas as pd
field = ['appln_id']
patents = 0
Category = ['Adapt','Build', 'Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Other-1', 'Other-3', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Other-4', 'Other-5', 'Nuclear', 'Non-Fossil', 'Enable','Other-6']
for c in Category:
    df = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Extras/'+ c + '_Residence_complete.csv', skipinitialspace=True, usecols=field, header=0)
    n = df['appln_id'].nunique()
    patents = patents + n

patents

We have retained 7,931,602 of 8,042,737 patents.
