==> Data Cleaning/7_Transforming to Panel Data.ipynb <==
import pandas as pd
import numpy as np
import time

### Creating a Panel format that includes 1. Total patents 2. Total Successful Patents 3. Successful Patents with Family size >=2 

# Category = ['1_ExTemp', '3_Storm', '4_Flood', '5_Landslide', '6_WaveAct', '7_Drought', '9_Wildfire', 'Solar-T', 'Wind', 'Nuclear', 'Hydrogen', 'Adapt','Build','Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'PV', 'Thermal-PV', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Other-1', 'Other-3', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Non-Fossil', 'Enable', 'Other-4', 'Other-5','Other-6']
Category = [ 'Other-5','Other-6']
# Category = [ 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Other-1', 'Other-3', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Nuclear', 'Non-Fossil', 'Enable', 'Other-4', 'Other-5','Other-6']
for c in Category:
    df1 = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Extras/' + c + '_everything.csv', skipinitialspace=True, header=0)
    df1.loc[df1.appln_filing_year == 9999] = np.NaN
    df1.dropna(subset=['appln_filing_year'], inplace=True)

    countries = pd.unique( pd.concat([df1['person_ctry_code'],df1['person_ctry_code.1'],df1['person_ctry_code.2'],df1['person_ctry_code.3'],df1['person_ctry_code.4']]))


    years = pd.unique(df1['appln_filing_year'])
    start_year = 1985
    end_year = years.max()
    start_time = time.time()
        
    print('Counting ', c)
    list_of_points = []
    for country in countries:
        patent_country_idx = ( ((df1['person_ctry_code'] == country) | (df1['person_ctry_code.1'] == country) | (df1['person_ctry_code.2'] == country) | (df1['person_ctry_code.3'] == country) | (df1['person_ctry_code.4'] == country)))
        K_stock = sum((df1['appln_filing_year'] <= start_year) & patent_country_idx &  (df1['granted'] == 'Y'))            
        for year in range(int(start_year), int(end_year) + 1):
            patent_country_year_idx = (df1['appln_filing_year'] == year) & patent_country_idx
            count = sum(patent_country_year_idx)
            granted_count = sum(patent_country_year_idx & (df1['granted'] == 'Y'))
            granted_family_ge_2 = sum(patent_country_year_idx & (df1['granted'] == 'Y') & (df1['docdb_family_size'] >=2 ))
            K_stock = (1 - 0.15) * K_stock + granted_count
            list_of_points.append({'country': country, 'year': year, 'total_count': count, 'granted_count': granted_count , 'granted_family_ge_2_count': granted_family_ge_2, 'K_stock':K_stock })
    df_panel = pd.DataFrame.from_records(list_of_points)
    df_panel.to_csv('Volumes/NJ_4TB/PATSAT/MyFiles/Extras/' + c + '_panel.csv', encoding='utf-8', index=False)
    del df1, df_panel
    print("Finished in ", time.time() - start_time)
