==> Data Cleaning/8_EM_DAT cleaning.ipynb <==
### Here I have built a different dataset for each disaster. The format is panel, with country-year units. 

import pandas as pd
import numpy as np

df1 = pd.read_csv('/Volumes/NJ_4TB/EMDAT/EMDAT.csv', skipinitialspace=True, skiprows=6)

disasters = pd.unique(df1['Disaster Type'])
countries = pd.unique(df1['Country'])
print('# Disaster types: ', len(disasters) )
print('# Countries : ', len(countries) )
#NaNs in interested columns tobe considered 0
df1['Total Deaths'].fillna(0, inplace = True)
df1['Total Affected'].fillna(0, inplace = True)
df1["Total Damages, Adjusted ('000 US$)"].fillna(0, inplace = True)
start_year = 1984
end_year = 2020


for disaster in disasters:    
    print("Processing disaster: ", disaster)
    disaster_all_dataframes = []
    for country in countries:
        sub_index = (df1['Country'] == country) & (df1['Disaster Type'] == disaster)
        single_year_events = sub_index & (df1['End Year'] == df1['Start Year'])
        multi_year_events = np.where(sub_index & (df1['End Year'] != df1['Start Year']))[0]
        list_of_points = []
        for year in range(start_year, end_year + 1):
            indices = single_year_events & (df1['Year'] == year)
            total_deaths = sum(df1['Total Deaths'][indices])
            total_affected = sum(df1['Total Affected'][indices])
            total_damages_adj = sum(df1["Total Damages, Adjusted ('000 US$)"][indices])
            list_of_points.append({'country':country, 'year':year, 'total_deaths':total_deaths, 'total_affected':total_affected, 'total_damages_adj':total_damages_adj})
        disaster_df = pd.DataFrame.from_records(list_of_points)
        for event in multi_year_events:
            n_years = df1.iloc[event]['End Year'] - df1.iloc[event]['Start Year'] + 1.0
            damage = df1.iloc[event]["Total Damages, Adjusted ('000 US$)"].copy() / n_years
            affected = df1.iloc[event]["Total Affected"] / n_years
            deaths = df1.iloc[event]["Total Deaths"] / n_years
            for year in range( df1.iloc[event]['Start Year'],  df1.iloc[event]['End Year'] + 1):
                if np.where(disaster_df['year'] == year)[0].size>0:
                    r = np.where(disaster_df['year'] == year)[0][0]
                    disaster_df.at[r,'total_deaths'] += deaths
                    disaster_df.at[r,'total_affected'] +=  affected
                    disaster_df.at[r,'total_damages_adj'] += damage
        disaster_all_dataframes.append(disaster_df)
    disaster_all_countries = pd.concat(disaster_all_dataframes)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(11).sum()).add_prefix('r10_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(10).sum()).add_prefix('r9_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(9).sum()).add_prefix('r8_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(8).sum()).add_prefix('r7_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(7).sum()).add_prefix('r6_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(6).sum()).add_prefix('r5_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(5).sum()).add_prefix('r4_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(4).sum()).add_prefix('r3_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(3).sum()).add_prefix('r2_')],axis=1)
    disaster_all_countries = pd.concat([disaster_all_countries, disaster_all_countries.groupby('country')[['total_deaths','total_affected','total_damages_adj']].transform(lambda s: s.rolling(2).sum()).add_prefix('r1_')],axis=1)
    disaster_all_countries.to_csv('/Volumes/NJ_4TB/EMDAT/' + disaster + '_EMDAT_panel.csv', encoding='utf-8', index=False) 
#     disaster_all_countries.to_csv(disaster + '_EMDAT_panel.csv', encoding='utf-8', index=False) 

disaster_all_countries[disaster_all_countries['country']=='India']

df=pd.read_csv('/Volumes/NJ_4TB/EMDAT/Storm_EMDAT_panel.csv', skipinitialspace=True)
df[df['country']=='India']
