==> Data Cleaning/10_Combine_EMDAT_PATSAT.ipynb <==
import pandas as pd
import numpy as np
import re
from itertools import chain
from pathlib import Path

#### Map file names

emdat_patsat_names = {'1_ExTemp':'Extreme temperature ', '3_Storm':'Storm', '4_Flood':'Flood', '5_Landslide':'Landslide', '7_Drought':'Drought', '9_Wildfire':'Wildfire'}

#### Map countries

countries = []
for c in emdat_patsat_names.keys():
    df = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/' + c + '_panel.csv', skipinitialspace=True, header=0)
    countries = countries + pd.unique(df['country']).tolist()
patsat_unique_countries = list(set(countries))
print("# unique countries in PATSAT : ", len(patsat_unique_countries))

countries = []
for c in emdat_patsat_names.values():
    df = pd.read_csv('/Volumes/NJ_4TB/EMDAT/'+ c + '_EMDAT_panel.csv', skipinitialspace=True, header=0)
    countries = countries + pd.unique(df['country']).tolist()
emdat_unique_countries = list(set(countries))
print("# unique countries in EMDAT : ", len(emdat_unique_countries))

patsat_unique_countries

r = re.compile(".*saud", re.IGNORECASE)
print(list(filter(r.match, emdat_unique_countries)))

emdat_patsat_countries={ 'KR':'Korea (the Republic of)',
                         'MX':'Mexico',
                         'EG':'Egypt',
                         'JP':'Japan',
                         'NO':'Norway',
                         'CA':'Canada',
                         'SI':'Slovenia',
                         'DK':'Denmark',
                         'TH':'Thailand',
                         'FI':'Finland',
                         'HU':'Hungary',
                         'IS':'Iceland',
                         'IN':'India',
                         'US':'United States of America (the)',
                         'CZ':['Czech Republic (the)', 'Czechoslovakia','Slovakia'],
                         'CL':'Chile',
                         'PL':'Poland',
                         'LV':'Latvia',
                         'KY':'Cayman Islands (the)',
                         'PH':'Philippines (the)',
                         'DE':['Germany Dem Rep', 'Germany', 'Germany Fed Rep'],#also DD
                         'SG':'Singapore',
                         'CH':'Switzerland',
                         'LT':'Lithuania',
                         'CY':'Cyprus',
                         'RU':'Russian Federation (the)',
                         'UA':'Ukraine',
                         'IT':'Italy',
                         'AU':'Australia',
                         'NZ':'New Zealand',
                         'BE':'Belgium',
                         'QA':'Qatar',
                         'MY':'Malaysia',
                         'BM':'Bermuda',
                         'AT':'Austria',
                         'SU':'Soviet Union',
                         'GR':'Greece',
                         'MT':'Malta',
                         'IE':'Ireland',
                         'TR':'Turkey',
                         'ES':'Spain',
                         'NL':'Netherlands (the)',
                         'IL':'Israel',
                         'MA':'Morocco',
                         'AE':'United Arab Emirates (the)',
                         'CU':'Cuba',
                         'SE':'Sweden',
                         'CN':'China',
                         'FR':'France',
                         'LU':'Luxembourg',
                         'PA':'Panama',
                         'BG':'Bulgaria',
                         'ZA':'South Africa',
                         'TW':'Taiwan (Province of China)',
                         'AR':'Argentina',
                         'BR':'Brazil',
                         'GB':'United Kingdom of Great Britain and Northern Ireland (the)',
                         'SA':'Saudi Arabia'
                       }

emdat_patsat_map = {}
for key, val in emdat_patsat_countries.items():
    if key == 'CZ' or key == 'DE':
        for i, coun in enumerate(val):
            emdat_patsat_map[coun] = key + str(i)
    else:
        emdat_patsat_map[val] = key
emdat_patsat_map

### Combine PATSAT, EMDAT
#### also, combine DE and DD in PATSAT

# for patsat, emdat in emdat_patsat_names.items():
#     df_patsat =  pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/' + patsat + '_panel.csv', skipinitialspace=True, header=0)
#     df_emdat = pd.read_csv(emdat + '_EMDAT_panel.csv', skipinitialspace=True, header=0)
#     ## combine DE with DD
#     if np.where(df_patsat['country'] == 'DD')[0].size>0:
#         for year in range( 1985,  2020 + 1):
#             dd_year = np.where((df_patsat['year'] == year) & (df_patsat['country'] == 'DD'))
#             if dd_year[0].size>0:
#                 dd = dd_year[0][0]
#                 de = np.where((df_patsat['year'] == year) & (df_patsat['country'] == 'DE') )[0][0]
#                 df_patsat.at[de,'total_count'] += df_patsat.at[dd,'total_count']
#                 df_patsat.at[de,'granted_count'] +=  df_patsat.at[dd,'granted_count']
#                 df_patsat.at[de,'granted_family_ge_2_count'] += df_patsat.at[dd,'granted_family_ge_2_count']
#     df_patsat.drop(df_patsat[df_patsat['country'] == 'LI'].index, inplace=True)
#     df_patsat.drop(df_patsat[df_patsat['country'] == 'VG'].index, inplace=True)
#     df_patsat.drop(df_patsat[df_patsat['country'] == 'DD'].index, inplace=True)
#     df_patsat.drop(df_patsat[df_patsat['country'] == 'FO'].index, inplace=True)
#     df_patsat["total_deaths"] = 0
#     df_patsat["total_affected"] = 0
#     df_patsat["total_damages_adj"] = 0
#     for country in pd.unique(df_patsat.country):
#         if country == 'DE' or country == 'CZ':
#             emdat_countries = emdat_patsat_countries[country]
#         else:
#             emdat_countries = [emdat_patsat_countries[country]]
        
#         for ec in emdat_countries:
#             for year in range( 1985,  2020 + 1):
#                 pt_year = np.where((df_patsat['year'] == year) & (df_patsat['country'] == country))
#                 if pt_year[0].size>0:
#                     pt = pt_year[0][0]
#                     em = np.where((df_emdat['year'] == year) & (df_emdat['country'] == ec) )[0][0]
#                     df_patsat.at[pt,'total_deaths'] = df_emdat.at[em,'total_deaths']
#                     df_patsat.at[pt,'total_affected'] =  df_emdat.at[em,'total_affected']
#                     df_patsat.at[pt,'total_damages_adj'] = df_emdat.at[em,'total_damages_adj']
#     if patsat=='1_ExTemp':
#         print(pd.unique(df_patsat['country']))
        
#     print('Combined ' + patsat)    
#     df_patsat.to_csv(patsat + '_panel_combined.csv', encoding='utf-8', index=False)
        

for patsat, emdat in emdat_patsat_names.items():
    df_patsat =  pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/' + patsat + '_panel.csv', skipinitialspace=True, header=0)
    df_emdat = pd.read_csv('/Volumes/NJ_4TB/EMDAT/' + emdat + '_EMDAT_panel.csv', skipinitialspace=True, header=0)
    if np.where(df_patsat['country'] == 'DD')[0].size>0:
        for year in range( 1985,  2020 + 1):
            dd_year = np.where((df_patsat['year'] == year) & (df_patsat['country'] == 'DD'))
            if dd_year[0].size>0:
                dd = dd_year[0][0]
                de = np.where((df_patsat['year'] == year) & (df_patsat['country'] == 'DE') )[0][0]
                df_patsat.at[de,'total_count'] += df_patsat.at[dd,'total_count']
                df_patsat.at[de,'granted_count'] +=  df_patsat.at[dd,'granted_count']
                df_patsat.at[de,'granted_family_ge_2_count'] += df_patsat.at[dd,'granted_family_ge_2_count']
                df_patsat.at[de,'K_stock'] += df_patsat.at[dd,'K_stock']
    df_patsat.drop(df_patsat[df_patsat['country'] == 'LI'].index, inplace=True)
    df_patsat.drop(df_patsat[df_patsat['country'] == 'VG'].index, inplace=True)
    df_patsat.drop(df_patsat[df_patsat['country'] == 'DD'].index, inplace=True)
    df_patsat.drop(df_patsat[df_patsat['country'] == 'FO'].index, inplace=True)
    df_emdat['country'] = df_emdat['country'].map(emdat_patsat_map).fillna(df_emdat['country'])
    list_columns=df_emdat.columns.to_list()
    list_columns.remove('year')
    list_columns.remove('country')
    df_de = df_emdat[df_emdat['country']=='DE0'][list_columns].reset_index().drop(columns=['index']) + df_emdat[df_emdat['country']=='DE1'][list_columns].reset_index().drop(columns=['index']) + df_emdat[df_emdat['country']=='DE2'][list_columns].reset_index().drop(columns=['index'])
    df_de['country']='DE'
    df_de['year']=df_emdat[df_emdat['country']=='DE0']['year'].reset_index().drop(columns=['index'])
    df_de.drop(0,inplace=True)

    df_cz = df_emdat[df_emdat['country']=='CZ0'][list_columns].reset_index().drop(columns=['index']) + df_emdat[df_emdat['country']=='CZ1'][list_columns].reset_index().drop(columns=['index']) + df_emdat[df_emdat['country']=='CZ2'][list_columns].reset_index().drop(columns=['index'])
    df_cz['country']='CZ'
    df_cz['year']=df_emdat[df_emdat['country']=='CZ0']['year'].reset_index().drop(columns=['index'])
    df_cz.drop(0,inplace=True)
    df_emdat = pd.concat([df_de,df_cz,df_emdat])
    df_combined = df_patsat.merge(df_emdat,on=['country','year'], how='inner')
    df_combined.to_csv('/Volumes/NJ_4TB/Final/' + patsat + '_panel_combined.csv', encoding='utf-8', index=False)
    print('Combined ' + patsat)  

### Combine green technologies PATSAT data with the combined data above

Categories = ['Build','Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Nuclear', 'Non-Fossil', 'Enable',  'Other-1', 'Other-3', 'Other-4']#,'Other-6', 'Other-5']
#Categories = ['Other-4','Other-5', 'Other-6']
df_controls =pd.read_csv('/Volumes/NJ_4TB/Final/controls_processed.csv', skipinitialspace=True, header=0)
for category in Categories:
    print("Now processing: "+ category)
    try:
        df_cat = pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/' + category + '_panel.csv', skipinitialspace=True, header=0)
    except pd.errors.EmptyDataError:
        df_cat = pd.DataFrame()
    if df_cat.empty:
        print('Skipped '+ category)
        continue
    if np.where(df_cat['country'] == 'DD')[0].size>0:
        for year in range( 1985,  2020 + 1):
            dd_year = np.where((df_cat['year'] == year) & (df_cat['country'] == 'DD'))
            if dd_year[0].size>0:
                dd = dd_year[0][0]
                de = np.where((df_cat['year'] == year) & (df_cat['country'] == 'DE') )[0][0]
                df_cat.at[de,'total_count'] += df_cat.at[dd,'total_count']
                df_cat.at[de,'granted_count'] +=  df_cat.at[dd,'granted_count']
                df_cat.at[de,'granted_family_ge_2_count'] += df_cat.at[dd,'granted_family_ge_2_count']
                df_cat.at[de,'K_stock'] += df_cat.at[dd,'K_stock']
    df_cat.drop(df_cat[df_cat['country'] == 'LI'].index, inplace=True)
    df_cat.drop(df_cat[df_cat['country'] == 'VG'].index, inplace=True)
    df_cat.drop(df_cat[df_cat['country'] == 'DD'].index, inplace=True)
    df_cat.drop(df_cat[df_cat['country'] == 'FO'].index, inplace=True)
    df_cat = pd.concat([pd.get_dummies(df_cat['year']), df_cat],axis=1)
    df_cat.loc[:, 'K_stock'] = np.log(1 + df_cat['K_stock'])
    df_cat.rename(columns={'total_count':category+'_total_count', 'granted_count':category +'_granted_count', 'granted_family_ge_2_count':category+'_granted_family_ge_2_count', 'K_stock':category+'_K_stock'}, inplace=True)
    df_cat = df_cat.merge(df_controls, on=['country','year'], how='left')#NaNs if controls are not present 
    for adaptive in emdat_patsat_names.keys():        
        df_adapt = pd.read_csv('/Volumes/NJ_4TB/Final/' + adaptive + '_panel_combined.csv', skipinitialspace=True, header=0)
        df_adapt.loc[:, 'K_stock'] = np.log(1 + df_adapt['K_stock'])
        df_adapt.rename(columns={'total_count':adaptive+'_total_count', 'granted_count':adaptive+'_granted_count', 'granted_family_ge_2_count':adaptive+'_granted_family_ge_2_count', 'K_stock':adaptive+'_K_stock'}, inplace=True)
        df_final=df_cat.merge(df_adapt, on=['country','year'], how='inner')
        df_final.loc[:, 'gdp_per_capita'] = np.log(df_final['gdp_per_capita'])
        df_final.loc[:, 'control_total_patents'] = np.log(df_final['control_total_patents'])
        df_final.drop(df_final[df_final['year'] >2018].index, inplace=True)
        df_final.drop(labels =2019, axis =1, inplace=True, errors='ignore')
        df_final.drop(labels =2020, axis =1, inplace=True, errors='ignore')
        #  print('Combined ' + category + ' with '+ adaptive)    
        directory = '/Volumes/NJ_4TB/Final/Adaptive_Mitigative/'# + category + '_' + adaptive + '/'
        Path(directory).mkdir(parents=True, exist_ok=True)
        df_final.to_csv( directory + category + '_' + adaptive + '_panel_combined.csv', encoding='utf-8', index=False)
        

pd.read_csv('/Volumes/NJ_4TB/PATSAT/MyFiles/Other-6_panel.csv', skipinitialspace=True, header=0)

#extracting chosen country samples
import pandas as pd
import numpy as np
Categories = ['Build','Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Nuclear', 'Non-Fossil', 'Enable',  'Other-1', 'Other-3']#, 'Other-5','Other-6', 'Other-4']
emdat_patsat_names = {'1_ExTemp':'Extreme temperature ', '3_Storm':'Storm', '4_Flood':'Flood', '5_Landslide':'Landslide', '7_Drought':'Drought', '9_Wildfire':'Wildfire'}
countries = {'Country': ['KR','MX','EG','JP','NO','CA','SI','DK','TH','FI','HU','IS','IN','US','CZ','CL','PL','LV','KY','PH','DE','SG','CH','LT', 'CY',                          
'RU','UA','IT','AU','NZ','BE','QA','MY','BM','AT','SU','GR','MT','IE','TR','ES','NL','IL','MA','AE','CU', 'SE',                          
'CN','FR','LU','PA','BG','ZA','TW','AR','BR','GB','SA']}                         
sample_list = pd.DataFrame(countries)
for a in emdat_patsat_names:
    for c in Categories:
        df = pd.read_csv('/Volumes/NJ_4TB/Final/Adaptive_Mitigative/'+ c + '_'+ a + '_panel_combined.csv', skipinitialspace=True, header=0)
        sample = pd.Series(df['country'].unique())
        sample_list[c + '_' + a] = sample_list.iloc[:, 0].isin(sample)
sample_list.to_csv('/Volumes/NJ_4TB/Final/Adaptive_Mitigative/sample_list.csv', index=False)
sample_list


#Filter Sample by Country
import pandas as pd
import numpy as np
Categories = ['Build','Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Nuclear', 'Non-Fossil', 'Enable',  'Other-1', 'Other-3']#, 'Other-5','Other-6', 'Other-4']
emdat_patsat_names = {'1_ExTemp':'Extreme temperature ', '3_Storm':'Storm', '4_Flood':'Flood', '5_Landslide':'Landslide', '7_Drought':'Drought', '9_Wildfire':'Wildfire'}
#countries = {'KR', 'JP', 'NO', 'CA', 'DK', 'HU', 'IN', 'US', 'DE', 'CH', 'RU', 'IT', 'AU', 'BE', 'AT', 'ES', 'NL', 'IL', 'SE', 'CN', 'FR', 'TW', 'BR', 'GB'}
for a in emdat_patsat_names:
    for c in Categories:
        df = pd.read_csv('/Volumes/NJ_4TB/Final/Adaptive_Mitigative/'+ c + '_'+ a + '_panel_combined.csv', skipinitialspace=True, header=0)
        df = df[df['country'].str.contains('(KR|JP|NO|CA|DK|HU|IN|US|DE|CH|RU|IT|AU|BE|AT|ES|NL|IL|SE|CN|FR|TW|BR|GB)', regex=True)]
        df.to_csv('/Volumes/NJ_4TB/Final/Adaptive_Mitigative/'+ c + '_'+ a + '_panel_24countries.csv', index=False)
