==> Data Cleaning/11_Merging all Green tecnologies .ipynb <==
import pandas as pd
import numpy as np
Categories = ['Sequester', 'ICT', 'Geo', 'Hydro', 'Ocean', 'Solar-T', 'PV', 'Thermal-PV', 'Wind', 'Combust', 'Fusion', 'Fission', 'Electric', 'Biofuel', 'Waste', 'Store- B', 'Store - C', 'Store- T', 'Store- M', 'Hydrogen', 'Processing', 'Transport', 'Wastewater', 'SmrtGrds', 'Nuclear', 'Non-Fossil', 'Enable',  'Other-1', 'Other-3']#, 'Other-5','Other-6', 'Other-4']
emdat_patsat_names = {'1_ExTemp', '3_Storm', '4_Flood', '5_Landslide', '7_Drought', '9_Wildfire'}
for a in emdat_patsat_names:
    df_merged = pd.read_csv('/Volumes/NJ_4TB/Final/Adaptive_Mitigative/Build' + '_'+ a + '_panel_combined.csv', skipinitialspace=True, header=0)
    total_count = ['Build_total_count']
    granted_count = ['Build_granted_count']
    ge2 = ['Build_granted_family_ge_2_count']
    kstock = ['Build_K_stock']
    for c in Categories:
        df2 = pd.read_csv('/Volumes/NJ_4TB/Final/Adaptive_Mitigative/'+ c + '_'+ a + '_panel_combined.csv', skipinitialspace=True, header=0)
        df_merged = pd.merge(df_merged, df2[['country','year',c+'_total_count',c+'_granted_count',c+'_granted_family_ge_2_count',c+'_K_stock']],on=['country','year'], how = 'left')
        total_count.append(c+'_total_count')
        granted_count.append(c+ '_granted_count')
        ge2.append(c+'_granted_family_ge_2_count')
        kstock.append(c+'_K_stock')
    df_merged.fillna(0, inplace=True)
    df_merged['Total_count'] =  df_merged[total_count].sum(axis=1)
    df_merged['Granted_count'] =  df_merged[granted_count].sum(axis=1)
    df_merged['Granted_family_ge_2_count'] =  df_merged[ge2].sum(axis=1)
    df_merged['K_Stock'] =  df_merged[kstock].sum(axis=1)
    df_merged.drop(total_count+granted_count+ge2+kstock, axis=1, inplace=True)
    df_merged.to_csv('/Volumes/NJ_4TB/Final/Sum_'+ a + '_panel_combined.csv', index=False)
